Refactor the energy_processor tests into their own file
